Make the og:image tag override-able
Despite the GDS comment that says their og:image is just a fall-back, it
turns out a few scrapers and sharers use either ALL or the LAST
og:image, so a fall-back is actually an override. I've moved it to the
override-able head template section so our pages provide JUST our icons.

Issue: Scottish-Natural-Heritage/Licensing#381