💥 Move from `node-sass` to `sass`
This is a breaking-ish change, but in a good way! `node-sass` has been
deprecated and we should all be using `sass`. This new version has its
own package/module resolver so we can do away with
`node-sass-package-importer`. It does mean that we need to remove the
`~` prefix from our imports and instead specify `--load-path`s when we
call `sass` to build the `.scss`s into `.css`s.